Создан декоратор, извлекающий данные пользователя из request.headers. Добавлено ограничение возможности удаления пользователей: только админ может удалять других пользователей, остальные пользователи имеют право удалять только себя. Создана миграция, добавляющая каскадное удаление данных из таблицы tokens при удалении связанных данных из таблицы users.

+++ apps/server/src/user/user.service.ts
@@ -1,7 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { UserRepository } from '@server/user/user.repository';
 import { SignUpUserDto } from '@server/auth/dto';
 import { GetUsersDto, UpdateUserDto } from '@server/user/dto';
+import type { JwtPayload } from '@server/auth/interfaces';
+import { UserRoleEnum } from '@server/user/enums/user-role.enum';
 
 @Injectable()
 export class UserService {
@@ -23,7 +25,10 @@ export class UserService {
     return this.userRepository.updateEntity(id, updateUserDto);
   }
 
-  remove(id: number) {
+  remove(id: number, user: JwtPayload) {
+    if (user.sub !== id && !user.roles.includes(UserRoleEnum.ADMIN)) {
+      throw new ForbiddenException('You are not allowed to delete this user');
+    }
     return this.userRepository.removeEntity(id);
   }
 }
